+++ substrate/frame/election-provider-multi-phase/test-staking-e2e/src/lib.rs
@@ -152,10 +152,12 @@ fn mass_slash_doesnt_enter_emergency_phase() {
 		// active set should stay the same before and after the slash
 		assert_eq!(active_set_size_before_slash, active_set_size_after_slash);
 
-		// Slashed validators are disabled up to a threshold
-		slashed.truncate(pallet_staking::UpToThresholdDisablingStrategy::<
-			SLASHING_DISABLING_FACTOR,
-		>::disable_threshold(active_set_size_after_slash));
+		// Slashed validators are disabled up to a limit
+		slashed.truncate(
+			pallet_staking::UpToLimitDisablingStrategy::<SLASHING_DISABLING_FACTOR>::disable_limit(
+				active_set_size_after_slash,
+			),
+		);
 
 		// Find the indices of the disabled validators
 		let active_set = Session::validators();